Restructure state management in QuestionPage and its components
* Question.jsx -> QuestionCard.jsx
* Remove QuestionButton component for now, it made the state passing pretty confusing
* Pull the question state upwards to the QuestionPage
* Make MUI buttons to integrate with react-router, and use it for the question navigation
* Show how many questions is left at the bottom
