+++ frontend/src/theme.jsx
@@ -1,4 +1,5 @@
 import { createTheme } from '@mui/material'
+import { LinkBehavior } from './components/LinkBehavior'
 
 export const theme = createTheme({
     palette: {
@@ -6,4 +7,12 @@ export const theme = createTheme({
             main: '#1976d2' /* MUI default */,
         },
     },
+    components: {
+        MuiLink: {
+            defaultProps: { component: LinkBehavior }
+        },
+        MuiButtonBase: {
+            defaultProps: { LinkComponent: LinkBehavior }
+        }
+    }
 })
